+++ contracts/toCKB-typescript/src/utils/mod.rs
@@ -1,3 +1,5 @@
+pub mod common;
 pub mod tools;
+
 pub use tockb_types as types;
 pub use tockb_types::config;
